requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.1
